requirements.txt
gunicorn
selenium
psycopg2-binary
pycrypto
flask
flask-wtf
flask-sqlalchemy
flask-migrate
